simplify handleNestedLinks and extract pure functions to derivedLinksTypes
